modules/api/location_ingestor/requirements.txt
Flask==1.1.1
Flask_SocketIO==5.3.6
kafka_python==2.0.2
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
gunicorn==21.2.0
Jinja2==2.11.2
Werkzeug==0.16.1
itsdangerous==1.1.0
MarkupSafe==1.1.1
